+++ include/mgard-x/Linearization/LevelLinearizer.hpp
@@ -257,9 +257,9 @@ public:
     tbz = R;
     tby = C;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
     for (int d = D - 4; d >= 0; d--) {
       gridx *= v.shape(d);
     }
